+++ cluster/constants.py
@@ -4,6 +4,7 @@ JSON_SETTINGS_FILE = "settings.json"
 JOB_INFO_FILE = "job_info.csv"
 
 STATUS_PICKLE_FILE = "status.pickle"
+SUBMISSION_HOOK_STATS_FILE = "submission_hook_stats.pickle"
 FULL_DF_FILE = "all_data.csv"
 REDUCED_DF_FILE = "reduced_data.csv"
 STD_ENDING = "__std"