+++ cluster/latex_utils.py
@@ -1,6 +1,7 @@
 import datetime
 import logging
 import os
+import typing
 from abc import ABC, abstractmethod
 from shutil import copyfile
 from subprocess import PIPE, CalledProcessError, run
@@ -79,7 +80,17 @@ class LatexFile(object):
         )
         self.sections.append(section(name, content))
 
-    def produce_pdf(self, output_file):
+    def produce_pdf(self, output_file: str) -> None:
+        """Construct the LaTeX file and generate a PDF from it (using pdflatex).
+
+        In case pdflatex fails with an error, the LaTeX file is saved to ``{{
+        output_file }}.tex`` and the log to `{{ output_file }}.log`` for offline
+        debugging.  If there is no error, those files are automatically deleted when
+        finished.
+
+        Args:
+            output_file: Path to which the PDF file is written.
+        """
         logger = logging.getLogger("cluster_utils")
         full_content = "\n".join(self.sections)
         title_str = LATEX_TITLE.format(self.title)
@@ -99,7 +110,7 @@ class LatexFile(object):
                     check=True,
                     stdout=PIPE,
                 )
-            except CalledProcessError as e:
+            except CalledProcessError:
                 # save the log and tex for debugging
                 logger.error("pdflatex failed.  Save log and tex file for debugging.")
                 latex_log_file = os.path.join(tmpdir, "latex.log")
@@ -107,7 +118,7 @@ class LatexFile(object):
                 copyfile(latex_log_file, output_file + ".log")
 
                 # re-raise the error
-                raise e
+                raise
 
             logger.info(f"pdflatex call produced output file {output_file}")
             output_tmp = os.path.join(tmpdir, "latex.pdf")
@@ -116,6 +127,22 @@ class LatexFile(object):
 
 def latex_format(string):
     return string.replace("_", "-")
+
+
+class StaticSectionGenerator:
+    """
+    Section generator (for use with :class:`SectionHook`) that returns a static value.
+    """
+
+    def __init__(self, value: typing.Any) -> None:
+        """
+        Args:
+            value:  Value that will be returned when calling the instance.
+        """
+        self.value = value
+
+    def __call__(self, df, path_to_results, filename_generator):
+        return self.value
 
 
 class SectionHook(ABC):